tests: remove jest-serial-runner
This was causing an entire separate copy of Jest 24 to be
installed. Mixing pieces of Jest 24 and other versions finally
broke with Jest 27.

In its place, switch to `--runInBand` as describe as equivalent in
jest-serial-runner's documentation. Actually that flag is already
present in the integration test that seems to need it (as the tests
send a sequence of requests to the test device) already uses this
flag. I'm adding it to the unit tests too, to minimize the impact
of this change. Feel free to remove that if you can determine that
it's not necessary.